requirements.txt
Django==4.1.4
python-decouple==3.6
djangorestframework==3.14.0
Pillow==9.3.0
pyotp==2.8.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
django-minio-backend==3.3.2
drf-spectacular==0.25.1
